+++ Jarvis.Framework/Support/GuaranteedDeliveryBroadcastBlock.cs
@@ -8,20 +8,30 @@ using System.Threading.Tasks.Dataflow;
 
 namespace Jarvis.Framework.Kernel.Support
 {
+    /// <summary>
+    /// An helper that creates a broadcast block that will send the message to all the targets
+    /// using TPL dataflow
+    /// </summary>
     public static class GuaranteedDeliveryBroadcastBlock
     {
+        /// <summary>
+        /// Create a broadcast using a list of TargetBlocks.
+        /// </summary>
+        /// <exception cref="JarvisFrameworkEngineException"></exception>
         public static ActionBlock<T> Create<T>(
             IEnumerable<ITargetBlock<T>> targets,
             String commitPollingClientId,
             Int32 boundedCapacity,
-            Int32 secondsToWaitBeforeThrowError = 10)
+            Int32 secondsToWaitBeforeThrowError = 10,
+            string meterName = null)
         {
             var options = new ExecutionDataflowBlockOptions();
             if (boundedCapacity > 0)
             {
                 options.BoundedCapacity = boundedCapacity;
             }
-            var meter = JarvisFrameworkMetric.Meter($"GuaranteedDeliveryBroadcastBlock-{commitPollingClientId}", Unit.Items, TimeUnit.Seconds);
+            meterName = meterName ?? $"GuaranteedDeliveryBroadcastBlock-{commitPollingClientId}";
+            var meter = JarvisFrameworkMetric.Meter(meterName, Unit.Items, TimeUnit.Seconds);
             var actionBlock = new ActionBlock<T>(
                 async item =>
                 {
@@ -31,10 +41,10 @@ namespace Jarvis.Framework.Kernel.Support
                         while (!(await target.SendAsync(item).ConfigureAwait(false)))
                         {
                             //message was not sent to the target, we need to wait a little bit and retry, if we fail too many times we raise an exception.
-                            await Task.Delay(1000);
+                            await Task.Delay(5000);
                             if (errorCount > secondsToWaitBeforeThrowError)
                             {
-                                throw new JarvisFrameworkEngineException("GuaranteedDeliveryBroadcastBlock: Unable to send message to a target id " + commitPollingClientId + "  of type " + item.GetType());
+                                throw new JarvisFrameworkEngineException("GuaranteedDeliveryBroadcastBlock" + meterName + ": Unable to send message to a target id " + commitPollingClientId + "  of type " + item.GetType());
                             }
                             errorCount++;
                         }